% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_functions.r
\name{dctdistance}
\alias{dctdistance}
\title{Get distance between two DCT sets}
\usage{
dctdistance(x, y)
}
\arguments{
\item{x, y}{data frame of DCT coefficients}
}
\value{
vector of euclidean distance between rows
}
\description{
This function returns a vector of distances between the DCT coefficient sets
  provided to it. The vector index is equivalent to the row index of the
  input data frames. The distance is calculated as an n-dimensional
  Euclidean distance
}
\examples{
x = data.frame(replicate(10,rnorm(10,500,50))) # Create arbitrary data
y = data.frame(replicate(10,rnorm(10,500,50)))
xcoeffs = getdct(x) # Get DCT coefficients
ycoeffs = getdct(y)
dctdistance(xcoeffs,ycoeffs)
}
\seealso{
\link[voweltrajectories]{getdct} whose output this function takes as input.
}
