% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_functions.r
\name{dcterror}
\alias{dcterror}
\title{Get error of a DCT model}
\usage{
dcterror(n, data)
}
\arguments{
\item{n}{integer specifying number of DCT coefficients to use}

\item{data}{data frame of signal observations as columns}
}
\value{
A numeric representing the sum of squared error for a DCT model
}
\description{
This function returns the sum of squared errors for a dataset modeled
  using a given number of DCT coefficients.
}
\examples{
x = data.frame(replicate(10,rnorm(10,500,50))) # Create arbitrary data
dcterror(3,x) # Get error of model with 3 DCT coefficients
}
\seealso{
\link[voweltrajectories]{dctprediction} which provides model predictions.
}
