% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_functions.r
\name{dctprediction}
\alias{dctprediction}
\title{Get predictions of a DCT model}
\usage{
dctprediction(n, data)
}
\arguments{
\item{n}{integer specifying number of DCT coefficients to use}

\item{data}{data frame of signal observations as columns}
}
\value{
A matrix containing the predictions of a DCT model
}
\description{
This function returns the predicted values for a dataset modeled
  using a given number of DCT coefficients.
}
\examples{
x = data.frame(replicate(10,rnorm(10,500,50))) # Create arbitrary data
dctprediction(3,x) # Get error of model with 3 DCT coefficients
}
\seealso{
\link[dtt]{dct} which this function wraps.
}
