% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_functions.r
\name{getdct}
\alias{getdct}
\title{Get DCT coefficients}
\usage{
getdct(data, n = 0)
}
\arguments{
\item{data}{data frame with signal observations as columns}

\item{n}{integer specifying number of DCT coefficients to return, defaults to all}
}
\value{
A data frame of DCT coefficients as columns.
}
\description{
This function returns a data frame with columns representing DCT coefficients
  for the data input. Input data must have one row for each data series
  and each row must containe only the data. This function currently does
  not handle data frames which also have metadata, so users need to handle
  the addition of metadata on their end. The function guarantees that
  output rows match input rows.
}
\examples{
x = data.frame(replicate(10,rnorm(10,500,50))) # Create arbitrary data
getdct(x,n=3) # Get first 3 DCT coefficients
}
\seealso{
\link[dtt]{dct} which this function wraps.
}
