% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_functions.r
\name{getelbow}
\alias{getelbow}
\title{Get elbow point of model errors}
\usage{
getelbow(errs, series = FALSE)
}
\arguments{
\item{errs}{vector of sum of squared errors}

\item{series}{logical; if TRUE returns a vector of second derivative estimates}
}
\value{
A numeric representing the elbow point of a data series
}
\description{
This function returns the index in the given error series which the series has an
  elbow point. The elbow point is defined as the maximum absolute estimated second
  derivative. IF the error series starts at the error of a model using one DCT
  coefficient, the output represents the best number of coefficients to use in
  an analysis. If you specify a different starting point
  in \link[voweltrajectories]{getprederr}, this will not be the case; it is the user's
  responsibility to know the relationship between error series indices and
  number of DCT coefficients used.
}
\details{
Optionally, it can return the entire estimated second derivative, though this is
  deprecated and will be separated in future versions.
}
\examples{
x = data.frame(replicate(10,rnorm(10,500,50))) # Create arbitrary data
err = getprederr(x)
getelbow(err)
}
\seealso{
\link[voweltrajectories]{getprederr} whose output this function takes as input.
}
