% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_functions.r
\name{getprederr}
\alias{getprederr}
\title{Get errors of various numbers of DCT coefficients}
\usage{
getprederr(data, start = 1, stop = NA)
}
\arguments{
\item{data}{data frame of signal observations as columns}

\item{start}{integer specifying the minimum number of coefficients to use; defaults to 1}

\item{stop}{integer specifying the maximum number of coefficients to usel defaults to all}
}
\value{
A vector of sum of squares for each number of DCT coefficients
}
\description{
This function returns a vector of the sum of squared errors for each
  number of DCT coefficients. By with no other arguments, it returns
  the sum of squares for all numbers of DCT coefficients, that is, from
  1 to the number of observations (columns). The function takes
  optional start and stop values to limit the number of coefficients
  evaluated which can be useful for signals with high sample rates.
}
\examples{
x = data.frame(replicate(10,rnorm(10,500,50))) # Create arbitrary data
getprederr(x)
}
\seealso{
\link[voweltrajectories]{dcterror} which this function calls.
}
