% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{VowelB}
\alias{VowelB}
\title{Hypothetical formant data}
\format{
a dataframe with 10 rows and 10 columns
}
\usage{
VowelB
}
\description{
Each column is a hypothetical speaker from A to J.
Each row is a formant measurement at intervals 1:10
}
\keyword{datasets}
